chore: rename quota-operator to platform-service-quota
