+++ PROJECT
@@ -5,8 +5,8 @@
 domain: openmcp.cloud
 layout:
 - go.kubebuilder.io/v4
-projectName: quota-operator
-repo: github.com/openmcp-project/quota-operator
+projectName: platform-service-quota
+repo: github.com/openmcp-project/platform-service-quota
 resources:
 - api:
     crdVersion: v1
@@ -14,6 +14,6 @@ resources:
   controller: true
   domain: openmcp.cloud
   kind: QuotaIncrease
-  path: github.com/openmcp-project/quota-operator/api/v1alpha1
+  path: github.com/openmcp-project/platform-service-quota/api/v1alpha1
   version: v1alpha1
 version: "3"
