+++ cmd/platform-service-quota/app/run.go
@@ -26,9 +26,9 @@ import (
 	openmcpconst "github.com/openmcp-project/openmcp-operator/api/constants"
 	"github.com/openmcp-project/openmcp-operator/lib/clusteraccess"
 
-	providerscheme "github.com/openmcp-project/quota-operator/api/install"
-	quotav1alpha1 "github.com/openmcp-project/quota-operator/api/v1alpha1"
-	"github.com/openmcp-project/quota-operator/internal/controller/quota"
+	providerscheme "github.com/openmcp-project/platform-service-quota/api/install"
+	quotav1alpha1 "github.com/openmcp-project/platform-service-quota/api/v1alpha1"
+	"github.com/openmcp-project/platform-service-quota/internal/controller/quota"
 )
 
 var setupLog logging.Logger
