+++ go.mod
@@ -1,15 +1,15 @@
-module github.com/openmcp-project/quota-operator
+module github.com/openmcp-project/platform-service-quota
 
 go 1.26.2
 
-replace github.com/openmcp-project/quota-operator/api => ./api
+replace github.com/openmcp-project/platform-service-quota/api => ./api
 
 require (
 	github.com/onsi/ginkgo/v2 v2.28.1
 	github.com/onsi/gomega v1.39.1
 	github.com/openmcp-project/controller-utils v0.27.0
 	github.com/openmcp-project/openmcp-operator/lib v0.17.1
-	github.com/openmcp-project/quota-operator/api v0.18.0
+	github.com/openmcp-project/platform-service-quota/api v0.18.0
 	github.com/spf13/cobra v1.10.2
 	k8s.io/api v0.35.3
 	k8s.io/apimachinery v0.35.3
