+++ hack/environment.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 # this file is sourced automatically by the environment.sh script in the 'common' scripts submodule
 
-export MODULE_NAME="github.com/openmcp-project/quota-operator"
+export MODULE_NAME="github.com/openmcp-project/platform-service-quota"
 export NESTED_MODULES="api"
