+++ internal/controller/quota/controller_test.go
@@ -20,9 +20,9 @@ import (
 	openmcpctrlutil "github.com/openmcp-project/controller-utils/pkg/controller"
 	testutils "github.com/openmcp-project/controller-utils/pkg/testing"
 
-	quotainstall "github.com/openmcp-project/quota-operator/api/install"
-	quotav1alpha1 "github.com/openmcp-project/quota-operator/api/v1alpha1"
-	quotacontroller "github.com/openmcp-project/quota-operator/internal/controller/quota"
+	quotainstall "github.com/openmcp-project/platform-service-quota/api/install"
+	quotav1alpha1 "github.com/openmcp-project/platform-service-quota/api/v1alpha1"
+	quotacontroller "github.com/openmcp-project/platform-service-quota/internal/controller/quota"
 )
 
 const (
